move change stream runner code into spec and test files to match the defined classes
